backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
